requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
cryptography==46.0.3
dj-database-url==3.1.2
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.13.0
django-autocomplete-light==3.12.1
django-cloudinary-storage==0.3.0
django-extensions==4.1
django-filter==25.2
django-phonenumber-field==8.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
Markdown==3.10
phonenumbers==9.0.17
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
